Use a 0-sized ArrayList instead of a LinkedList
The list serves as the initial value of the path and is only copied into
other ArrayLists (in getPath()), never modified. It's more consistent to
also use an ArrayList here, since seeing LinkedList raises questions as to
why it was chosen. Sizing it at 0 avoids garbage, and depending on the JDK
can also kick off optimizations such as using a constant, empty array
(Oracle JDK 8, for example).

The JIT compiler might also generate slightly more efficient native code for
getPath(), as a single implementation of List will be observed, instead of
2: all method calls are monomorphic instead of bimorphic, so the code is
simpler.